Switched to Filters, Changed Interval, Added Page Check
Changed everything over to use filters instead of actions. Added a check
for specific admin pages to display on with default as all admin pages.
And made interval the amount of delay in days before showing a notice
instead of time before auto-dismiss. Thus notices will only be dismissed
by clicking the dismiss instead of based on time as well.