Concept Update
Shows how the extended class can add a function to update backend
information when the notice is dismissed. The function name is
irrelevant but I just used the same as the default. The If statement in
the extended class needs to ensure the ignore isset and that it is equal
to the slug used by the notice to prevent interference with other
notices.