Ability to specify a "name-addr" element for the CC, BCC, Reply-To fields
The specs on Email Message Formatting allow for the CC, BCC, Reply-To
fields to contain a full name-addr element. (i.e. John Smith
<john.smith@example.org>).  This patch allows for the full format to
specified in the field and to be used.

Example: Set "Reply-To" to "{field:textbox_41} <{field:email_42}>" under
Advanced where {field:textbox_41} is "Name" and  {field:email_42} is an
"Email"